move TableCard logic to tab-conent.jsx
