+++ mailpoet/assets/js/src/newsletters/standard/tab-content.jsx
@@ -7,14 +7,11 @@ import { MailPoet } from 'mailpoet';
 import PropTypes from 'prop-types';
 import { useLocation, useParams } from 'react-router-dom';
 
-import { TableCard,TablePlaceholder } from '@woocommerce/components';
-import { useSelect, useDispatch } from '@wordpress/data';
+import { TableCard } from '@woocommerce/components';
+import { useSelect } from '@wordpress/data';
 import { storeName } from '../store/constants';
-import { QueueStatus } from 'newsletters/listings/queue-status';
-import { Statistics } from 'newsletters/listings/statistics.jsx';
-import { ErrorBoundary, withBoundary } from '../../common';
-import { SegmentTags, FilterSegmentTag } from '../../common/tag/tags';
 
+import { getRow } from './get-row';
 
 const mailpoetTrackingEnabled = MailPoet.trackingConfig.emailTrackingEnabled;
 const gutenbergEmailEditorSupported = MailPoet.FeaturesController.isSupported('gutenberg_email_editor');
@@ -50,30 +47,14 @@ export const NEWSLETTER_STANDARD_HEADERS = [
 
 
 
-const selectSegementsFromIds = (segment_ids, segments) => {
-  return segment_ids.map(segment_id => segments.find(segment => segment.id === segment_id));
-}
+
 
 
 // TODO receive meta from the store
-function transformToTableCardRows(standard, segments, meta = { mta_log: {}, current_time: '' }) {
-  return standard.map(newsletter => {
+function transformToTableCardRows(standard, segments, ) {
+  const meta = { mta_log: {}, current_time: '' };
 
-    return [
-      { display: newsletter.subject, value: newsletter.subject },
-      { display: <QueueStatus newsletter={newsletter} mailerLog={meta.mta_log} />, value: newsletter.status },
-      { display: <div
-          className="column mailpoet-hide-on-mobile"
-          data-colname={__('Lists', 'mailpoet')}>
-          <ErrorBoundary>
-            <SegmentTags segments={selectSegementsFromIds(newsletter.segment_ids, segments)} dimension="large" />
-            <FilterSegmentTag newsletter={newsletter} dimension="large" />
-          </ErrorBoundary>
-        </div>, value: newsletter.segment_ids },
-      { display: <Statistics newsletter={newsletter} currentTime={meta.current_time} />, value: newsletter.statistics },
-      { display: newsletter.sent_at ? newsletter.sent_at : 'Not sent', value: newsletter.sent_at || '' },
-    ];
-  });
+  return standard.map(newsletter =>  getRow("tab-name", newsletter, segments, meta) );
 }
 
 
@@ -105,7 +86,6 @@ export function StandardTabContent() {
             query={ { page: 2 } }
             rowsPerPage={ 7 }
             totalRows={ 10 }
-         //   summary={ summary }
         />
         </div>
